# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreateNoteRequest:
    properties:
      content:
        type: string
      room_id:
        type: integer
      title:
        type: string
    type: object
...
  dtos.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  dtos.NoteResponse:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      room_id:
        type: integer
      room_name:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_email:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
...
  dtos.UpdateNoteRequest:
    properties:
      content:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for room reservation management with user authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Room Reservation API
  version: "1.0"
paths:
...
  /notes:
    post:
      consumes:
      - application/json
      description: Create a new note in a room
      parameters:
      - description: Note creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dtos.NoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create note
      tags:
      - notes
  /notes/{note_id}:
    delete:
      consumes:
      - application/json
      description: Delete note (only by note creator)
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete note
      tags:
      - notes
    get:
      consumes:
      - application/json
      description: Retrieve detailed note information
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.NoteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get note by ID
      tags:
      - notes
    put:
      consumes:
      - application/json
      description: Update note information (only by note creator)
      parameters:
      - description: Note ID
        in: path
        name: note_id
        required: true
        type: integer
      - description: Note update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update note
      tags:
      - notes
  /notes/my-notes:
    get:
      consumes:
      - application/json
      description: Get all notes created by the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.NoteResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my notes
      tags:
      - notes
  /notes/room/{room_id}:
    get:
      consumes:
      - application/json
      description: Get all notes in a specific room
      parameters:
      - description: Room ID
        in: path
        name: room_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.NoteResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dtos.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get notes by room
      tags:
      - notes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"